==> workshop_materials/bike_demand_forecasting/02_data_preparation.ipynb <==
# 🚀 Module 2: Data Preparations

# Install requirements
!pip install -r requirements.txt

## 📦 Import Required Libraries

Before we proceed with training and tracking our machine learning model, we need to import the necessary libraries.


# Import necessary modules
from sklearn.model_selection import train_test_split

import pandas as pd
import os

## 📥 Load the Processed Dataset

We'll start by loading the processed dataset for January 2011, which was prepared in the data exploration phase.

This dataset contains cleaned and feature-engineered data and will be used as the reference dataset for drift and performance comparison.

We use `pandas` to read the CSV file and inspect the first few rows.

# Load the training data
data_path = "./data/processed/"

# Read both CSV files
# TODO: YOU SHOULD SET THE PATH TO CLEAND DATASET FOR THE FIRST TWO MONTH
data_01 = pd.read_csv(data_path + 'DATA_MONTH_1')
data_02 = pd.read_csv(data_path + 'DATA_MONTH_2')
# data_03 = pd.read_csv(data_path + 'data_2011_03.csv')

# Concatenate the datasets
# input_data_df = pd.concat([data_01, data_02, data_03], ignore_index=True)
input_data_df = pd.concat([data_01, data_02], ignore_index=True)

input_data_df.head()

## 🔧 Feature Selection
In this section, we define the features used for model training by separating them into numerical and categorical variables:

- Numerical Features:

        ['temp', 'atemp', 'humidity', 'windspeed', 'hour', 'weekday']

These features represent continuous or ordered values, such as temperature and time-related attributes.

- Categorical Features:

        ['season', 'holiday', 'workingday', 'weathersit']

These features represent discrete categories or binary flags that indicate specific conditions or periods.

This separation is essential for preprocessing, allowing us to apply appropriate transformations to each feature type.

numerical_features=['temp', 'atemp', 'humidity', 'windspeed', 'hour', 'weekday']
# TODO: YOU SHOULD PASTE THE CATEGORICAL FEATURES HERE
categorical_features='PLEASE INSERT CATEGORICAL FEATURES HERE'  # Replace with actual categorical features

## ✂️ Define Features and Target Variable

We prepare the dataset for training by selecting input features and the target variable:

- Input Features (X_input): A combination of both numerical and categorical features defined earlier.

- Target Variable (y_input): The target for prediction is the count column, representing the total number of bike rentals.

Next, we perform a train-test split using an 70/30 ratio to evaluate model performance on unseen data. The random_state=42 ensures reproducibility of results.

X_input = input_data_df[numerical_features + categorical_features]
y_input = input_data_df["count"]

# Train-test split
# TODO: YOU SHOULD PASTE THE FUNCTION FOR DATASET SPLITTING TO TRAIN AND TEST
X_train, X_test, y_train, y_test = "ADD TRAIN_TEST_SPLIT FUNCTION HERE"  # Replace with actual train-test split function

Displaying the first few rows of the **train feature set** to verify the structure and content:

X_train.head()

Displaying the first few rows of the **test feature set** to verify the structure and content:

X_test.head()

## 💾 Save Train and Test Datasets

We'll save the train and test datasets for later use in model training.

data_dir ="./data"
split_dir = os.path.join(data_dir, "split")

# Create directories if they don't exist
os.makedirs(split_dir, exist_ok=True)

# Save the split datasets to CSV
X_train.to_csv(f"{split_dir}/X_train.csv", index=False)
X_test.to_csv(f"{split_dir}/X_test.csv", index=False)
y_train.to_csv(f"{split_dir}/y_train.csv", index=False)
y_test.to_csv(f"{split_dir}/y_test.csv", index=False)


#### Congratulations! You have completed all the steps in task 2 (`Data Preparations for Model Training`). 
#### Please go back to the instructions on the GitHub-Pages.
